# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BR Tours and Travels API
  version: 0.3.0
  description: >-
    Express business API for the public website, authenticated administration,
    media, enquiries and notification operations. Booking requests are leads,
    not payments, issued inventory or guaranteed reservations.
servers:
  - url: /api/v1
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Verify credentials and rotate into an opaque MySQL session
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      requestBody:
        required: true
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/LoginInput" } },
          }
      responses:
        "200": { description: HttpOnly session cookie and explicit admin DTO }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /auth/logout:
    post:
      tags: [Authentication]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      responses: { "204": { description: Session revoked and cookie cleared } }
  /auth/me:
    get:
      tags: [Authentication]
      security: [{ cookieAuth: [] }]
      responses: { "200": { description: Current explicit admin identity DTO } }
  /auth/profile:
    put:
      tags: [Authentication]
      summary: Update the current admin profile
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      responses: { "200": { description: Updated explicit admin identity DTO } }
  /auth/change-password:
    post:
      tags: [Authentication]
      summary: Change the current password and revoke every session
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      responses:
        { "204": { description: Password changed and session cookie cleared } }
  /auth/forgot-password:
    post:
      tags: [Authentication]
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      responses: { "202": { description: Enumeration-safe reset response } }
  /auth/reset-password:
    post:
      tags: [Authentication]
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      responses:
        {
          "200":
            { description: Single-use token consumed and sessions revoked },
        }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: br_admin_session
      description: Environment-configured opaque session cookie; never a browser-storage token.
  parameters:
...
    CsrfToken:
      in: header
      name: x-csrf-token
      required: true
      schema: { type: string }
...
  schemas:
    LoginInput:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, maxLength: 256 }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            fields:
              {
                type: object,
                additionalProperties: { type: array, items: { type: string } },
              }
            requestId: { type: string }
  responses:
    Error:
      description: Safe API error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }